requirements.txt
streamlit
streamlit-chat
pillow
requests
langchain
tensorflow
numpy
opencv-python
pypdf2
patchify